refactor(#209): 클래스 상세페이지 컴포넌트 분리
